# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spool API",
    "version": "1.0.0",
    "description": "The REST API of a Spool server: a self-hosted git forge that keeps every repository in object storage. Git smart-HTTP endpoints (clone/fetch/push at /{org}/{repo}.git) speak the standard git wire protocol and are not described here. Every repository is private to the organization that owns it: a request with no credential answers 401 on every repository, changeset, search and profile route, and a credential that may not read a repository gets the same 404 a repository that does not exist gets. The exceptions are the sign-in routes under /v1/auth, the health and metrics probes, the inbound webhook receiver, the runner routes (which take their own credentials) and the signed CDN pack URL. Errors return JSON {\"error\": \"…\"} with the matching status. Optimistic-concurrency conflicts return 409 with the current tip."
  },
  "servers": [
    {
      "url": "https://spool.example.com",
      "description": "Your server; set by STRATUM_PUBLIC_URL."
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token of the form weft_<id>_<secret>. Scopes: org:admin, org:read, repo:read, repo:write. A token acts only in the organization it was minted in: presented anywhere else it has no access, and a repository there answers 404. Repo-bound tokens are restricted to one repository and cannot search."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "stratum_session",
        "description": "Browser session from POST /v1/auth/login. An alternative to a bearer token for the dashboard; a bearer token wins when both are presented. A session spans every organization its person belongs to, which is why anything that crosses two namespaces — forking, opening a change from a fork — takes one. Routes under /v1/auth use this scheme."
      }
    },
...
  },
  "paths": {
...
    "/v1/users/{handle}": {
      "get": {
        "summary": "A person's profile",
        "description": "Display name, bio, location, company, pronouns, account kind (human or agent) and links: the directory entry for a person on this server. Needs a signed-in caller — a session or a personal token; with none the answer is 401. It never carries an email address and never lists repositories. A disabled account has no profile and answers 404; its namespace stays claimed so nobody can move in behind them.",
        "responses": {
          "200": {
            "description": "{ handle, name, display_name, bio, location, company, pronouns, kind, profile_repo, created_at, links: [{ label, url }] }"
          },
          "404": {
            "description": "No such handle, or the account is disabled"
          },
          "401": {
            "description": "Nobody is signed in"
          }
        }
      },
      "patch": {
        "summary": "Edit your own profile",
        "description": "Every field is optional: an absent key leaves it alone, an explicit null clears it. `links` replaces the whole list (at most 5) and every url must be http or https — a profile link is rendered as an anchor on a page other people load, so `javascript:` and `data:` are refused. Nobody may patch anybody else's profile.",
        "responses": {
          "200": {
            "description": "The profile as it now reads"
          },
          "400": {
            "description": "A field is over its cap, holds a control character, or a link is not http(s)"
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "This is somebody else's account"
          },
          "404": {
            "description": "No such handle"
          }
        }
      }
    },
    "/v1/users/{handle}/emails": {
      "get": {
        "summary": "The addresses on your own account, proved or not",
        "description": "Readable only by the account itself, never by another. This is the authorship-linkage surface — the set that decides which commits count as this person's work — so showing it to anybody else would hand an impersonator the exact string to put in `git config user.email`, and a spammer a mailbox.",
        "responses": {
          "200": {
            "description": "{ emails: [{ address, verified_at, private, primary, created_at }] }"
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "This is somebody else's account"
          },
          "404": {
            "description": "No such handle"
          }
        }
      },
      "post": {
        "summary": "Claim an address and mail it a confirmation link",
        "description": "Body: { email }. The claim is worth nothing on its own: the row lands unproved, and only a verified address may ever carry authorship. The link goes to the mailbox and never into the response — a claim confirmed over the API would prove nothing. It is bound both to the address and to this account, so it cannot be spent against a different mailbox or from a different session. At most 10 addresses per account.",
        "responses": {
          "202": {
            "description": "{ address, status, detail }"
          },
          "400": {
            "description": "Not an address, or the account is at its address limit"
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "This is somebody else's account"
          },
          "409": {
            "description": "Already yours, or already in use. Which account holds it is never disclosed."
          }
        }
      }
    },
    "/v1/users/{handle}/emails/verify": {
      "post": {
        "summary": "Spend a confirmation link and mark that one address proved",
        "description": "Body: { token }. Every wrong shape — malformed, unknown, expired, already spent, or issued to another account — answers the same 404, so a link cannot be used to ask which addresses are registered here.",
        "responses": {
          "200": {
            "description": "{ address, verified: true }"
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "This is somebody else's account"
          },
          "404": {
            "description": "This confirmation link is not valid any more"
          }
        }
      }
    },
    "/v1/users/{handle}/emails/{email}": {
      "delete": {
        "summary": "Drop an address from your account",
        "description": "The address you sign in with is refused — an account with no reachable address is one nobody can recover — and so is an address that is not yours, with the same answer.",
        "responses": {
          "204": {
            "description": "Removed"
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "This is somebody else's account"
          },
          "404": {
            "description": "No such address on this account, or it is the address you sign in with"
          }
        }
      }
    },
    "/v1/orgs/{org}/profile": {
      "get": {
        "summary": "An organization's profile",
        "description": "Display name, description, location, website and contact address. For anybody in the organization, by session or by token. An org that has never been edited reads exactly like one whose fields were cleared — there is no second “no profile yet” state for a page to handle.",
        "responses": {
          "200": {
            "description": "{ org, display_name, description, location, website, contact_email }"
          },
          "401": {
            "description": "No credential, or one that does not authenticate"
          },
          "404": {
            "description": "No such org, or the caller is not in it — the two are not told apart"
          }
        }
      },
      "patch": {
        "summary": "Edit an organization's profile (org:admin)",
        "description": "Optional fields; an explicit null clears one. `website` must be http or https and `contact_email` must be an address, for the same reason the personal profile holds them to it: both are rendered on a page other people load.",
        "responses": {
          "200": {
            "description": "The profile as it now reads"
          },
          "400": {
            "description": "A field is over its cap, the website is not http(s), or the contact address is not an address"
          },
          "401": {
            "description": "No credential"
          },
          "403": {
            "description": "Needs org:admin"
          },
          "404": {
            "description": "No such org, or the caller is not a member"
          }
        }
      }
    },
...
  }
}